Fixed bug where longpressing on an interest in InterestView deleted it without confirmation.
git-svn-id: file:///home/gerard/programming/vidshare-svn/trunk@19 3e7b97c2-2e4e-4883-83ef-b2c1530b9b69

+++ VidshareHaggle/src/com/gm375/vidshare/InterestView.java
@@ -95,8 +95,8 @@ public class InterestView extends Activity {
     @Override
     public boolean onContextItemSelected(MenuItem item) {
         AdapterContextMenuInfo info = (AdapterContextMenuInfo) item.getMenuInfo();
-        
-        deleteInterest(info.position);
+        if (item.getTitle() == "Delete")
+            deleteInterest(info.position);
         
         return true;
     }
